use u64 rather than usize in file hashing paths (#485)
Using the file hashing components in WASM found a bug that using 32 but
usize causes errors when hashing the file.

This PR enforces the use of u64 everywhere along that path (and also
pins the wasm-bindgen version)

+++ deduplication/src/raw_xorb_data.rs
@@ -37,7 +37,7 @@ impl RawXorbData {
 
         debug_assert_le!(num_bytes, *MAX_XORB_BYTES);
 
-        let hash_and_len: Vec<_> = chunks.iter().map(|c| (c.hash, c.data.len())).collect();
+        let hash_and_len: Vec<_> = chunks.iter().map(|c| (c.hash, c.data.len() as u64)).collect();
         let cas_hash = xorb_hash(&hash_and_len);
 
         // Build the MDBCASInfo struct.
